+++ docs/astro.config.mjs
@@ -67,7 +67,6 @@ export default defineConfig({
 									{ label: 'Embedding into websites', slug: 'guides/embedding-maps-into-websites' },
 									{ label: 'AI Georeferencer', slug: 'guides/ai-georeferencer-for-aerial-imagery' },
 									{ label: 'Plotting vector data attributes', slug: 'guides/plot-and-chart-from-shapefile-fields' },
-									{ label: 'Segment Anything', slug: 'guides/segment-anything' },
 								],
 							},
 							{